refactor: lamp_graph_generator.js分割、heatmap/json_creatorコード品質改善
lamp_graph_generator.js(848行)をjs/lamp_graph/配下の4モジュールに分割し、
heatmap_generator.jsとjson_creator.jsのマジックナンバー定数化・関数分割・
未使用import削除を実施。
