+++ index.html
@@ -143,7 +143,7 @@
     <script type="module" src="./js/score_change_to_json.js"></script>
     <script type="module" src="./js/json_creator.js"></script>
     <script type="module" src="./js/heatmap_generator.js"></script>
-    <script type="module" src="./js/lamp_graph_generator.js"></script>
+    <script type="module" src="./js/lamp_graph/lamp_graph_main.js"></script>
 
     <script>
         document.addEventListener('DOMContentLoaded', function() {
